% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_val_N.R
\name{cross_val_N}
\alias{cross_val_N}
\title{Cross-validation for sNBLDA_glm model}
\usage{
cross_val_N(dat, y, nfold = 10, lambda1 = 0, ncpu = 4, flds)
}
\arguments{
\item{dat}{Data matrix of N by P where N is the number of subjects and P is the number of features}

\item{y}{class label (numeric)}

\item{nfold}{Number of folds to be considered}

\item{lambda1}{Tuning parameter vector}

\item{ncpu}{Number of CPUS to be parallel}

\item{flds}{list of fld sets}
}
\description{
Uses cross-validation to select the best tuning paramter
}
