% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_val_NC.R
\name{cross_val_NC}
\alias{cross_val_NC}
\title{Cross-validation for sNBLDA_glmsc model}
\usage{
cross_val_NC(
  dat,
  y,
  clinical,
  nfold = 10,
  lambda1 = 0,
  lambda2 = 0,
  ncpu = 4,
  flds
)
}
\arguments{
\item{dat}{Data matrix of N by P where N is the number of subjects and P is the number of features}

\item{y}{class label (numeric)}

\item{clinical}{matrix of clinical information}

\item{nfold}{Number of folds to be considered}

\item{lambda1}{Tuning parameter vector for genes}

\item{lambda2}{Tuning parameter vector for covariates}

\item{ncpu}{Number of CPUS to be parallel}

\item{flds}{list of fld sets}
}
\description{
Uses cross-validation to select the best tuning paramter combination
}
