% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sNBLDA_glm.R
\name{new_sBETA_hatN}
\alias{new_sBETA_hatN}
\title{sNBLDA_glm model}
\usage{
new_sBETA_hatN(datTr, y, lambda1)
}
\arguments{
\item{y}{class label (numeric)}

\item{lambda1}{Tuning parameter vector}

\item{dat}{Data matrix of N by P where N is the number of subjects and P is the number of features}
}
\description{
Builds the predictive model
}
