% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_sBETA_hatNC.R
\name{new_sBETA_hatNC}
\alias{new_sBETA_hatNC}
\title{sNBLDA_GLMsC}
\usage{
new_sBETA_hatNC(dat, y, lambda1, lambda2, clinical)
}
\arguments{
\item{dat}{Data matrix of N by P where N is the number of subjects and P is the number of features}

\item{y}{class label (numeric)}

\item{lambda1}{Tuning parameter vector for genes}

\item{lambda2}{Tuning parameter vector for covariates}

\item{clinical}{matrix of clinical information}
}
\description{
Fit the predictive model for sNBLDA_GLMsC
}
