% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_sNBLDA_glm.R
\name{Pred_function_N}
\alias{Pred_function_N}
\title{prediction for a new observation using sNBLDA_glm model}
\usage{
Pred_function_N(model, xte)
}
\arguments{
\item{model}{sNBLDA_glm model built using training data}

\item{xte}{Data matrix of N by P where N is the number of subjects and P is the number of features}
}
\description{
Predicts the class for a new observation
}
