% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pred_function_NC.R
\name{Pred_function_NC}
\alias{Pred_function_NC}
\title{prediction for a new observation using sNBLDA_glmsc model}
\usage{
Pred_function_NC(model, xte)
}
\arguments{
\item{model}{sNBLDA_glm model built using training data}

\item{xte}{Data matrix of N by P where N is the number of subjects and P is the number of features}
}
\description{
Predicts the class for a new observation
}
