% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulat_data.R
\name{Simulat_data}
\alias{Simulat_data}
\title{Simulate count data}
\usage{
Simulat_data(
  N_sample = 1100,
  pareto_shape = 8,
  pareto_scale = 10,
  lfc_mean = 0.5,
  cov_mean = 0,
  prob_inform = 0,
  N_train = 100
)
}
\arguments{
\item{N_sample}{Total number of samples to simulate}

\item{pareto_shape}{Value of the shape paramter}

\item{pareto_scale}{Value of the scale paramter}

\item{lfc_mean}{logfold change}

\item{cov_mean}{Mean effect size of the covariates}

\item{prob_inform}{Number of genes affected by clinical variables}

\item{N_train}{Number of training samples}
}
\description{
Simulate RNA-seq data with clinical variables
}
